flask/requirements.txt
bcrypt==4.0.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
defusedxml==0.7.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
python-gvm==23.11.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
